testing cml again

+++ libs/psbmpc_lib_gpu/tryms_matrix/eigen_interface.cuh
@@ -55,10 +55,10 @@ namespace CML
 		}
 	}
 
-	template<class T, int Max_Rows, int Max_Cols, typename Eigen_Type_T>
+	template<class T, size_t Max_Rows, size_t Max_Cols, typename Eigen_Type_T>
 	__host__ __device__ void assign_eigen_object(Pseudo_Dynamic_Matrix<T, Max_Rows, Max_Cols> &lhs, const Eigen_Type_T &rhs)
 	{
-		assert(Max_Rows >= rhs.rows() && Max_Cols >= rhs.cols());
+		assert((int)Max_Rows >= rhs.rows() && (int)Max_Cols >= rhs.cols());
 		int n_rows = rhs.rows(), n_cols = rhs.cols();
 		lhs.resize(n_rows, n_cols);
 		for (int i = 0; i < n_rows; i++)
@@ -70,11 +70,11 @@ namespace CML
 		}
 	}
 
-	template<class T, int Rows, int Cols, typename Eigen_Type_T>
+	template<class T, size_t Rows, size_t Cols, typename Eigen_Type_T>
 	__host__ __device__ void assign_eigen_object(Static_Matrix<T, Rows, Cols> &lhs, const Eigen_Type_T &rhs)
 	{
 		int n_rows = rhs.rows(), n_cols = rhs.cols();
-		assert(Rows == n_rows && Cols == n_cols);
+		assert((int)Rows == n_rows && (int)Cols == n_cols);
 
 		for (int i = 0; i < n_rows; i++)
 		{
@@ -107,7 +107,7 @@ namespace CML
 		}
 	}
 
-	template<class T, int Max_Rows, int Max_Cols, typename Eigen_Type_T>
+	template<class T, size_t Max_Rows, size_t Max_Cols, typename Eigen_Type_T>
 	__host__ __device__ void assign_cml_object(Eigen_Type_T &lhs, const Pseudo_Dynamic_Matrix<T, Max_Rows, Max_Cols> &rhs)
 	{
 		int n_rows = rhs.get_rows(), n_cols = rhs.get_cols();
@@ -121,7 +121,7 @@ namespace CML
 		}
 	}
 
-	template<class T, int Rows, int Cols, typename Eigen_Type_T>
+	template<class T, size_t Rows, size_t Cols, typename Eigen_Type_T>
 	__host__ __device__ void assign_cml_object(Eigen_Type_T &lhs, const Static_Matrix<T, Rows, Cols> &rhs)
 	{
 		lhs.resize(Rows, Cols);
